3/11/14
This commit is actually a culmination of the past week or so, just
working in short bursts. Combat is still random, which is infuriating,
and I was not able to complete the plot lines, but the two main paths
are somewhat flushed out, and there is still several ways to play the
game, so it is a success in that right. But I did get quite a bit of
work done with the events, but it was mostly writing, not so much
coding, which was expected.